Moved the settings into a class of their own.
